% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_games}
\alias{get_games}
\title{Returns a data frame with games from a season}
\usage{
get_games(season = season, team_id = NULL,
  access_token = get_shl_access_token())
}
\arguments{
\item{season}{Season to be used, i.e. '2015' for the SHL season 2015/2016}

\item{team_id}{If you only want the games for one team you can pass their team_id here, i.e. "FBK"}

\item{acces_token}{an access_token generated by get_shl_access_token()}
}
\description{
Returns a data frame with games from a season
}
\examples{
\dontrun{
get_teams()
}
}
