% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_standing.R
\name{get_team_standing}
\alias{get_team_standing}
\title{Returns a data.frame of teams and their statistics}
\usage{
get_team_standing(season = season, team_id = NULL,
  access_token = get_shl_access_token())
}
\arguments{
\item{season}{Season to be used, i.e. '2015' for the SHL season 2015/2016}

\item{team_id}{An ID for the team which the list should be based on, for example "DIF"}

\item{acces_token}{an access_token generated by get_shl_access_token()}
}
\description{
Returns a data.frame of teams and their statistics
}
\examples{
\dontrun{
get_teams()
}
}
