% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Returns a data frame with the teams currently active in SHL}
\usage{
get_teams(access_token = get_shl_access_token())
}
\arguments{
\item{acces_token}{an access_token generated by get_shl_access_token()}
}
\description{
Returns a data frame with the teams currently active in SHL
}
\examples{
\dontrun{
get_teams()
}
}
