% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top10_goalkeepers.R
\name{get_top10_goalkeeper_stats}
\alias{get_top10_goalkeeper_stats}
\title{Returns a data frame with top 10 goalkeeper statistics for a team or the whole league}
\usage{
get_top10_goalkeeper_stats(season = season, team_id = NULL,
  access_token = get_shl_access_token())
}
\arguments{
\item{season}{Season to be used, i.e. '2015' for the SHL season 2015/2016}

\item{team_id}{An ID for the team which the list should be based on, for example "DIF"}

\item{acces_token}{an access_token generated by get_shl_access_token()}
}
\description{
Returns a data frame with top 10 goalkeeper statistics for a team or the whole league
}
\examples{
\dontrun{
get_top10_goalkeepers(season = "2018", team_id = "DIF")
}
}
