% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top10_player_stats.R
\name{get_top10_player_stats}
\alias{get_top10_player_stats}
\title{Returns a data frame with top 10 player statistics for a team or the whole league}
\usage{
get_top10_player_stats(season = season, sort = "assists",
  team_id = NULL, access_token = get_shl_access_token())
}
\arguments{
\item{season}{Season to be used, i.e. '2015' for the SHL season 2015/2016}

\item{sort}{Order of which the top list should be sorted, could be "goals" or "assists", defaults to "assists"}

\item{team_id}{An ID for the team which the list should be based on, for example "DIF"}

\item{acces_token}{an access_token generated by get_shl_access_token()}
}
\description{
Returns a data frame with top 10 player statistics for a team or the whole league
}
\examples{
\dontrun{
get_teams()
}
}
