% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlr.R
\docType{package}
\name{shlr}
\alias{shlr}
\alias{shlr-package}
\title{\code{shlr} package}
\description{
A Quick and Easy Wrapper for data from SHL's Open API into R
}
\details{
See the README on
\href{https://github.com/filipwastberg/shlr#readme}{GitHub}
}
